refactor(git-panel): split GitPanel god component into modules + hooks
Break the 1091-line GitPanel.tsx into a GitPanel/ folder of focused,
reusable components and hooks. Behavior is unchanged (same JSX, class
names, and copy); verified with tsc, Biome, the test suite (304), and a
production build.

- Components: StatusHeader, SplitAction, CommitComposer, ChangesList,
  ActionBar, EmptyState, cards/ (PRCard, ClosedPRCard, ConflictCard,
  ChecksSection, CommentsSection), shared (Spinner, GitLink, ViewOnGitHub)
- Hooks: useGitPanelData (polling + derived reads), useCommitDraft,
  useTransientFeedback, useDelegationLifecycle, useGitActions (the dispatch
  table), useActionBarModel (split-button model)
- index.tsx is now a thin orchestrator wiring hooks to layout
- Dedup: one shared ViewOnGitHub button (was 3 copies) and a data-driven
  EmptyState (was 4 near-identical blocks)
