+++ src/components/RightPanel/GitPanel/cards/index.ts
@@ -1,0 +1,4 @@
+export { ChecksSection } from "./ChecksSection";
+export { CommentsSection } from "./CommentsSection";
+export { ConflictCard } from "./ConflictCard";
+export { ClosedPRCard, PRCard } from "./PRCard";
